Reorganize visitor.h
