+++ symengine/visitor.cpp
@@ -49,6 +49,36 @@ void postorder_traversal_stop(const Basic &b, StopVisitor &v)
     b.accept(v);
 }
 
+class HasSymbolVisitor : public BaseVisitor<HasSymbolVisitor, StopVisitor>
+{
+protected:
+    Ptr<const Symbol> x_;
+    bool has_;
+
+public:
+    HasSymbolVisitor(Ptr<const Symbol> x) : x_(x)
+    {
+    }
+
+    void bvisit(const Symbol &x)
+    {
+        if (x_->__eq__(x)) {
+            has_ = true;
+            stop_ = true;
+        }
+    }
+
+    void bvisit(const Basic &x){};
+
+    bool apply(const Basic &b)
+    {
+        has_ = false;
+        stop_ = false;
+        preorder_traversal_stop(b, *this);
+        return has_;
+    }
+};
+
 bool has_symbol(const Basic &b, const Symbol &x)
 {
     // We are breaking a rule when using ptrFromRef() here, but since
@@ -57,6 +87,75 @@ bool has_symbol(const Basic &b, const Symbol &x)
     HasSymbolVisitor v(ptrFromRef(x));
     return v.apply(b);
 }
+
+class CoeffVisitor : public BaseVisitor<CoeffVisitor, StopVisitor>
+{
+protected:
+    Ptr<const Symbol> x_;
+    Ptr<const Basic> n_;
+    RCP<const Basic> coeff_;
+
+public:
+    CoeffVisitor(Ptr<const Symbol> x, Ptr<const Basic> n) : x_(x), n_(n)
+    {
+    }
+
+    void bvisit(const Add &x)
+    {
+        umap_basic_num dict;
+        RCP<const Number> coef = zero;
+        for (auto &p : x.dict_) {
+            p.first->accept(*this);
+            if (neq(*coeff_, *zero)) {
+                Add::coef_dict_add_term(outArg(coef), dict, p.second, coeff_);
+            }
+        }
+        coeff_ = Add::from_dict(coef, std::move(dict));
+    }
+
+    void bvisit(const Mul &x)
+    {
+        for (auto &p : x.dict_) {
+            if (eq(*p.first, *x_) and eq(*p.second, *n_)) {
+                map_basic_basic dict = x.dict_;
+                dict.erase(p.first);
+                coeff_ = Mul::from_dict(x.coef_, std::move(dict));
+                return;
+            }
+        }
+        coeff_ = zero;
+    }
+
+    void bvisit(const Pow &x)
+    {
+        if (eq(*x.get_base(), *x_) and eq(*x.get_exp(), *n_)) {
+            coeff_ = one;
+        } else {
+            coeff_ = zero;
+        }
+    }
+
+    void bvisit(const Symbol &x)
+    {
+        if (eq(x, *x_) and eq(*one, *n_)) {
+            coeff_ = one;
+        } else {
+            coeff_ = zero;
+        }
+    }
+
+    void bvisit(const Basic &x)
+    {
+        coeff_ = zero;
+    }
+
+    RCP<const Basic> apply(const Basic &b)
+    {
+        coeff_ = zero;
+        b.accept(*this);
+        return coeff_;
+    }
+};
 
 RCP<const Basic> coeff(const Basic &b, const Basic &x, const Basic &n)
 {
